Consistent binary name for noderesourcetopology plugin
In the favour of consistency and making sure that the noderesourcetopology
plugin works with the existing deployment files we keep the binary name
same as the one used for other scheduler plugins i.e kube-scheduler.

In a lot of instances, the sys admin might want to provide additional
arguments or override the existing defaults in the deployment spec and
it is better to have the binary name same as the one used in case of
other scheduler plugins.

Signed-off-by: Swati Sehgal <swsehgal@redhat.com>